requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asgiref==3.11.1
<<<<<<< HEAD
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cryptography==46.0.5
defusedxml==0.7.1
distro==1.9.0
dj-database-url==3.1.2
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
kombu==5.6.2
oauthlib==3.3.1
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python3-openid==3.2.0
redis==7.2.1
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
=======
dj-database-url==3.1.2
Django==5.2.12
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
>>>>>>> main
